more things and stuff in azure

+++ remove-all-azure-resource-groups.ps1
@@ -2,6 +2,9 @@
 if ($sub.Id -eq 'b262856c-1efd-496c-9ce0-ed796a84b848')
 {
     Get-AzureRmResourceGroup | % {
-        Remove-AzureRmResourceGroup $_.ResourceGroupName -Force -AsJob
+        if ($_.ResourceGroupName -ne 'd00mFunction')
+        {
+            Remove-AzureRmResourceGroup $_.ResourceGroupName -Force -AsJob
+        }
     }
 }
